fix: make vest.reset restore initial state
